adds Label to RadioGrop
