+++ packages/radio-group/src/Radio/index.ts
@@ -1,0 +1,2 @@
+export { default as Radio } from './Radio';
+export type { RadioProps } from './Radio.types';
