+++ packages/radio-group/src/RadioGroup/RadioGroup.spec.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { axe } from 'jest-axe';
 
-import { RadioGroupProps, RadioProps } from './types';
-import { Radio, RadioGroup } from '.';
+import { RadioGroupProps, RadioProps } from '../types';
+import { Radio, RadioGroup } from '..';
 
 function WrappedRadio({ checked }: { checked: string }) {
   return (
