+++ packages/radio-group/src/RadioGroup/index.ts
@@ -1,0 +1,2 @@
+export { default as RadioGroup } from './RadioGroup';
+export type { RadioGroupProps } from './RadioGroup.types';
