+++ packages/radio-group/src/index.ts
@@ -1,5 +1,5 @@
-import Radio from './Radio';
-import RadioGroup from './RadioGroup';
+import { Radio, RadioProps } from './Radio';
+import { RadioGroup, RadioGroupProps } from './RadioGroup';
 import { Size } from './types';
 export { Radio, RadioGroup, Size };
-export type { RadioGroupProps, RadioProps } from './types';
+export type { RadioGroupProps, RadioProps };
