+++ packages/radio-group/tsconfig.json
@@ -25,6 +25,9 @@
       "path": "../tokens"
     },
     {
+      "path": "../typography"
+    },
+    {
       "path": "../leafygreen-provider"
     }
   ]
